MDEV-33616 Tests failing on macOS
These tests rely on THR_KEY_mysys but it is not initialized.  On
Linux, the corresponding thread variable is null, but on macOS it has a
nonzero value.  In all cases, initialize the variable explicitly.  We could
call MY_INIT, however some memory is leaked on test program exit in that
case.  Instead, factor just enough to be able to initialize THR_KEY_mysys
from the test code (as well as from the library code).